Add Windows Self Contained and .NET8.0 run configurations
Two new run configurations have been added for the project. The first one is "Windows Self Contained" which is a publish-to-folder type configuration. The other one is "Windows .NET8.0", which is also a publish-to-folder type configuration. Both are set up for the .NET 8.0 framework with the "Release" configuration and "Any CPU" platform.