Update README with new logger usage instructions
The README has been updated with detailed instructions on how to import and use the logger function in different ways. Both the HTML and ES6 module import methods have been included along with a note advising users to replace '{version}' with the actual version of the logger that they are using.